+++ minhas-financas-back-end/MinhasFinancas.Application/Interfaces/IMfScoreAuditoriaAppService.cs
@@ -5,5 +5,6 @@ namespace MinhasFinancas.Application.Interfaces
     public interface IMfScoreAuditoriaAppService
     {
         Task<RetornoGenerico> GerarPlanilhaAsync();
+        Task<RetornoGenerico> GerarPlanilhaAuditoriaHumanaAsync();
     }
 }